Split visualization functions into a separate file visualize.py
