+++ neossat/__init__.py
@@ -1,3 +1,4 @@
-from .neossatlib import *  # TODO might make this more complicated it we decide to split neossatlib a bit.
-from .improcess import *
 from .utils import *
+from .visualize import *
+from .neossatlib import *
+from .improcess import *
